requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
openai
tiktoken
requests
pydantic
